+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
      condition and the size of its results is invariant to all
      enclosing parallel loops.
 
+   * `futhark-dataset`: Add `-t`/`-type` option.  Useful for
+     inspecting data files.
+
 ### Removed
 
 ### Changed